requirements.txt
discord.py==2.3.2
google-generativeai==0.3.2
aiohttp==3.9.1
asyncio-throttle==1.0.2
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
aiofiles==23.2.1